[MOD] Numerous rewritings, optimizations, bug fixes

+++ basex-core/src/test/java/org/basex/server/LockingTest.java
@@ -30,9 +30,9 @@ public final class LockingTest extends SandboxTest {
   private static final int REPEAT = 1;
 
   /** Maximum sleep time in ms. */
-  private static final int SLEEP = 200;
+  private static final long SLEEP = 200L;
   /** Additional allowed holding time for client creation overhead, ... in ms. */
-  private static final int SYNC = 100;
+  private static final long SYNC = 100L;
 
   /** Test document. */
   private static final String DOC = "src/test/resources/test.xml";
